backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
celery>=5.4
redis>=5.0
django-celery-beat>=2.6
django-celery-results>=2.5
gunicorn>=21.2
whitenoise>=6.6
django-prometheus>=2.3
python-dotenv>=1.0
python-decouple>=3.8
